[common] Share a common implementation of ZpoolName
