+++ nexus/reconfigurator/execution/Cargo.toml
@@ -11,7 +11,6 @@ anyhow.workspace = true
 dns-service-client.workspace = true
 chrono.workspace = true
 futures.workspace = true
-illumos-utils.workspace = true
 internal-dns.workspace = true
 nexus-config.workspace = true
 nexus-db-model.workspace = true
