+++ nexus/reconfigurator/preparation/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 [dependencies]
 anyhow.workspace = true
 futures.workspace = true
-illumos-utils.workspace = true
 nexus-db-model.workspace = true
 nexus-db-queries.workspace = true
 nexus-types.workspace = true
